Update student_spec.rb
This method's spec is described as;

THE .ALL_STUDENTS_IN_GRADE_X METHOD
This is a class method that takes in an argument of grade for which to retrieve the roster. This method should return an array of all students for grade X.

The test seems to suggest that the method should be written to accept an argument (x) and then query the DB by grade (x). Instead it forces the student to read the spec and pass in "10" hardcoded as the query. Is this the correct behaviour?